[Fix] info page 컴포넌트 분리
